refactor: move ssao passes to its own file
