+++ systems/render-pipeline/post-processing-passes.js
@@ -1,254 +1,5 @@
-import { postProcessing as SHADERS } from "pex-shaders";
-import random from "pex-random";
-import { BlueNoiseGenerator } from "./blue-noise.js";
-
-// prettier-ignore
-const ssaoMixFlagDefinitions = [
-  [["options", "targets", "ssao.main"], "AO_TEXTURE", { type: "texture", uniform: "uSSAOTexture", requires: "USE_SSAO" }],
-  [["postProcessing", "ssao", "mix"], "", { uniform: "uSSAOMix", requires: "USE_SSAO" }],
-  [["postProcessing", "ssao", "type"], "USE_SSAO_GTAO", { compare: "gtao", requires: "USE_SSAO" }],
-  // [["postProcessing", "ssao", "type"], "USE_SSAO_SAO", { compare: "sao", requires: "USE_SSAO" }],
-  [["postProcessing", "ssao", "colorBounce"], "USE_SSAO_COLORS", { requires: "USE_SSAO_GTAO" }],
-]
-
-const ssao = ({
-  ctx,
-  resourceCache,
-  descriptors,
-  scale = 1,
-  saoMainTarget,
-  gtaoMainTarget,
-}) => [
-  // GTAO
-  {
-    name: "main",
-    frag: SHADERS.gtao.frag,
-    // blend: true,
-    // prettier-ignore
-    flagDefinitions: [
-        [["camera", "near"], "", { uniform: "uNear" }],
-        [["camera", "far"], "", { uniform: "uFar" }],
-        [["camera", "fov"], "", { uniform: "uFov" }],
-
-        [["postProcessing", "ssao"], "USE_SSAO"],
-
-        [["postProcessing", "ssao", "type"], "USE_SSAO_GTAO", { compare: "gtao", requires: "USE_SSAO" }],
-        [["postProcessing", "ssao", "slices"], "NUM_SLICES", { type: "value", requires: "USE_SSAO_GTAO" }],
-        [["postProcessing", "ssao", "samples"], "NUM_SAMPLES", { type: "value", requires: "USE_SSAO_GTAO" }],
-        [["postProcessing", "ssao", "intensity"], "", { uniform: "uIntensity", requires: "USE_SSAO_GTAO" }],
-        [["postProcessing", "ssao", "radius"], "", { uniform: "uRadius", requires: "USE_SSAO_GTAO" }],
-        [["postProcessing", "ssao", "brightness"], "", { uniform: "uBrightness", requires: "USE_SSAO_GTAO" }],
-        [["postProcessing", "ssao", "contrast"], "", { uniform: "uContrast", requires: "USE_SSAO_GTAO" }],
-        [["postProcessing", "ssao", "colorBounce"], "USE_COLOR_BOUNCE", { requires: "USE_SSAO_GTAO" }],
-        [["postProcessing", "ssao", "colorBounceIntensity"], "", { uniform: "uColorBounceIntensity", requires: "USE_SSAO_GTAO" }],
-        [["postProcessing", "ssao", "noiseTexture"], "USE_GTAO_NOISE_TEXTURE", { requires: "USE_SSAO_GTAO" }],
-        [["postProcessing", "ssao", "_gtaoNoiseTexture"], "NOISE_TEXTURE", { type: "texture", uniform: "uNoiseTexture", requires: "USE_GTAO_NOISE_TEXTURE" }],
-        [["postProcessing", "ssao", "_gtaoNoiseTexture", "width"], "", { uniform: "uNoiseTextureSize", requires: "USE_GTAO_NOISE_TEXTURE" }],
-      ],
-    disabled: ({ cameraEntity }) => {
-      const isEnabled = cameraEntity.postProcessing.ssao.type === "gtao";
-
-      if (
-        isEnabled &&
-        cameraEntity.postProcessing.ssao.noiseTexture &&
-        !cameraEntity.postProcessing.ssao._gtaoNoiseTexture
-      ) {
-        const generator = new BlueNoiseGenerator();
-        generator.size = 32;
-
-        const data = new Uint8Array(generator.size ** 2 * 4);
-        for (let i = 0, l = 4; i < l; i++) {
-          const result = generator.generate();
-          const bin = result.data;
-          const maxValue = result.maxValue;
-
-          for (let j = 0, l2 = bin.length; j < l2; j++) {
-            const value = 255 * (bin[j] / maxValue);
-            data[j * 4 + i] = value;
-          }
-        }
-        cameraEntity.postProcessing.ssao._gtaoNoiseTexture = ctx.texture2D({
-          width: generator.size,
-          height: generator.size,
-          data: data,
-          pixelFormat: ctx.PixelFormat.RGBA8,
-          encoding: ctx.Encoding.Linear,
-          wrap: ctx.Wrap.Repeat,
-          mag: ctx.Filter.Linear,
-          min: ctx.Filter.Linear,
-        });
-      }
-
-      return !isEnabled;
-    },
-    passDesc: () => ({
-      clearColor: [0, 0, 0, 1],
-    }),
-    target: ({ viewport }) => {
-      if (!gtaoMainTarget) {
-        gtaoMainTarget = ctx.texture2D({
-          ...descriptors.postProcessing.outputTextureDesc,
-          width: viewport[2] * scale,
-          height: viewport[3] * scale,
-        });
-      } else {
-        ctx.update(gtaoMainTarget, {
-          width: viewport[2] * scale,
-          height: viewport[3] * scale,
-        });
-      }
-      return gtaoMainTarget;
-    },
-    size: ({ viewport }) => [viewport[2] * scale, viewport[3] * scale],
-  },
-  // SAO
-  {
-    name: "main",
-    frag: SHADERS.sao.frag,
-    // prettier-ignore
-    flagDefinitions: [
-        [["camera", "near"], "", { uniform: "uNear" }],
-        [["camera", "far"], "", { uniform: "uFar" }],
-        [["camera", "fov"], "", { uniform: "uFov" }],
-
-        [["postProcessing", "ssao"], "USE_SSAO"],
-
-        [["postProcessing", "ssao", "type"], "USE_SSAO_SAO", { compare: "sao", requires: "USE_SSAO" }],
-        [["postProcessing", "ssao", "samples"], "NUM_SAMPLES", { type: "value", requires: "USE_SSAO_SAO" }],
-        [["postProcessing", "ssao", "spiralTurns"], "NUM_SPIRAL_TURNS", { type: "value", requires: "USE_SSAO_SAO" }],
-        [["postProcessing", "ssao", "intensity"], "", { uniform: "uIntensity", requires: "USE_SSAO_SAO" }],
-        [["postProcessing", "ssao", "bias"], "", { uniform: "uBias", requires: "USE_SSAO_SAO" }],
-        [["postProcessing", "ssao", "radius"], "", { uniform: "uRadius", requires: "USE_SSAO_SAO" }],
-        [["postProcessing", "ssao", "brightness"], "", { uniform: "uBrightness", requires: "USE_SSAO_SAO" }],
-        [["postProcessing", "ssao", "contrast"], "", { uniform: "uContrast", requires: "USE_SSAO_SAO" }],
-        [["postProcessing", "ssao", "noiseTexture"], "USE_SAO_NOISE_TEXTURE", { requires: "USE_SSAO_SAO" }],
-        [["postProcessing", "ssao", "_saoNoiseTexture"], "NOISE_TEXTURE", { type: "texture", uniform: "uNoiseTexture", requires: "USE_SAO_NOISE_TEXTURE" }],
-        [["postProcessing", "ssao", "_saoNoiseTexture", "width"], "", { uniform: "uNoiseTextureSize", requires: "USE_SAO_NOISE_TEXTURE" }],
-      ],
-    disabled: ({ cameraEntity }) => {
-      const isEnabled = cameraEntity.postProcessing.ssao.type === "sao";
-
-      if (
-        isEnabled &&
-        cameraEntity.postProcessing.ssao.noiseTexture &&
-        !cameraEntity.postProcessing.ssao._saoNoiseTexture
-      ) {
-        const localPRNG = random.create("0");
-
-        const size = 64;
-        const sizeSquared = size ** 2;
-        const channelSize = ctx.gl.RG ? 2 : 4;
-        const ssaoNoiseData = new Float32Array(sizeSquared * channelSize);
-        for (let i = 0; i < sizeSquared; i++) {
-          ssaoNoiseData[i * channelSize + 0] = localPRNG.float(-1, 1);
-          ssaoNoiseData[i * channelSize + 1] = localPRNG.float(-1, 1);
-          if (!ctx.gl.RG) {
-            ssaoNoiseData[i * channelSize + 2] = 0;
-            ssaoNoiseData[i * channelSize + 3] = 1;
-          }
-        }
-        cameraEntity.postProcessing.ssao._saoNoiseTexture = ctx.texture2D({
-          width: size,
-          height: size,
-          data: ssaoNoiseData,
-          pixelFormat: ctx.gl.RG
-            ? ctx.PixelFormat.RG32F
-            : ctx.PixelFormat.RGBA32F,
-          encoding: ctx.Encoding.Linear,
-          wrap: ctx.Wrap.Repeat,
-          mag: ctx.Filter.Linear,
-          min: ctx.Filter.Linear,
-        });
-      }
-
-      return !isEnabled;
-    },
-    passDesc: () => ({
-      clearColor: [0, 0, 0, 1],
-    }),
-    target: ({ viewport }) => {
-      if (!saoMainTarget) {
-        saoMainTarget = ctx.texture2D({
-          ...descriptors.postProcessing.outputTextureDesc,
-          // pixelFormat: ctx.PixelFormat.RGBA8,
-          pixelFormat: ctx.gl.RG ? ctx.PixelFormat.R8 : ctx.PixelFormat.RGBA8,
-          width: viewport[2] * scale,
-          height: viewport[3] * scale,
-        });
-      } else {
-        ctx.update(saoMainTarget, {
-          width: viewport[2] * scale,
-          height: viewport[3] * scale,
-        });
-      }
-      return saoMainTarget;
-    },
-    size: ({ viewport }) => [viewport[2] * scale, viewport[3] * scale],
-  },
-  {
-    name: "blurHorizontal",
-    frag: SHADERS.bilateralBlur.frag,
-    // prettier-ignore
-    flagDefinitions: [
-        [["camera", "near"], "", { uniform: "uNear" }],
-        [["camera", "far"], "", { uniform: "uFar" }],
-        [["postProcessing", "ssao", "blurSharpness"], "", { uniform: "uSharpness" }],
-      ],
-    disabled: ({ cameraEntity }) =>
-      cameraEntity.postProcessing.ssao.blurRadius === 0,
-    uniforms: ({ cameraEntity: entity }) => ({
-      uDirection: [entity.postProcessing.ssao.blurRadius, 0],
-    }),
-    passDesc: () => ({
-      clearColor: [0, 0, 0, 1],
-    }),
-    source: () => "ssao.main",
-    target: ({ viewport }) =>
-      resourceCache.texture2D({
-        ...descriptors.postProcessing.outputTextureDesc,
-        pixelFormat: ctx.PixelFormat.RGBA8,
-        width: viewport[2] * scale,
-        height: viewport[3] * scale,
-      }),
-    size: ({ viewport }) => [viewport[2] * scale, viewport[3] * scale],
-  },
-  {
-    name: "blurVertical",
-    frag: SHADERS.bilateralBlur.frag,
-    // prettier-ignore
-    flagDefinitions: [
-        [["camera", "near"], "", { uniform: "uNear" }],
-        [["camera", "far"], "", { uniform: "uFar" }],
-        [["postProcessing", "ssao", "blurSharpness"], "", { uniform: "uSharpness" }],
-      ],
-    disabled: ({ cameraEntity }) =>
-      cameraEntity.postProcessing.ssao.blurRadius === 0,
-    uniforms: ({ cameraEntity: entity }) => ({
-      uDirection: [0, entity.postProcessing.ssao.blurRadius],
-    }),
-    passDesc: () => ({
-      clearColor: [0, 0, 0, 1],
-    }),
-    source: () => "ssao.blurHorizontal",
-    target: () => "ssao.main",
-    size: ({ viewport }) => [viewport[2] * scale, viewport[3] * scale],
-  },
-  {
-    name: "mix",
-    frag: SHADERS.ssaoMix.frag,
-    // prettier-ignore
-    flagDefinitions: [
-      [["postProcessing", "ssao"], "USE_SSAO"],
-      ...ssaoMixFlagDefinitions,
-    ],
-    // If no dof, disable and do it in final
-    disabled: ({ cameraEntity }) => !cameraEntity.postProcessing.dof,
-    passDesc: () => ({
-      clearColor: [0, 0, 0, 1],
-    }),
-    size: ({ viewport }) => [viewport[2] * scale, viewport[3] * scale],
-  },
-];
+import { postProcessing as postprocessingShaders } from "pex-shaders";
+import { ssao, ssaoMixFlagDefinitions } from "./post-processing/ssao.js";
 
 const dof = ({ resourceCache, descriptors }) => [
   {
@@ -278,7 +29,7 @@ const dof = ({ resourceCache, descriptors }) => [
       [["postProcessing", "dof", "screenPoint"], "", { uniform: "uScreenPoint", requires: "USE_FOCUS_ON_SCREEN_POINT" }],
       [["postProcessing", "dof", "debug"], "USE_DEBUG", { requires: "USE_DOF" }],
     ],
-    frag: SHADERS.dof.frag,
+    frag: postprocessingShaders.dof.frag,
     target: ({ viewport }) =>
       resourceCache.texture2D({
         ...descriptors.postProcessing.outputTextureDesc,
@@ -293,7 +44,7 @@ const dof = ({ resourceCache, descriptors }) => [
 const bloom = ({ ctx, resourceCache, descriptors, bloomLevels = 9 }) => [
   {
     name: "threshold",
-    frag: SHADERS.threshold.frag,
+    frag: postprocessingShaders.threshold.frag,
     // prettier-ignore
     flagDefinitions: [
       [["camera", "exposure"], "", { uniform: "uExposure" }],
@@ -313,7 +64,7 @@ const bloom = ({ ctx, resourceCache, descriptors, bloomLevels = 9 }) => [
   // TODO: custom downsample levels based on viewport size
   ...Array.from({ length: bloomLevels }, (_, k) => k).map((i) => ({
     name: `downsample[${i}]`,
-    frag: SHADERS.downsample.frag,
+    frag: postprocessingShaders.downsample.frag,
     // prettier-ignore
     flagDefinitions: [
       [["postProcessing", "bloom"], "USE_DOWN_SAMPLE"],
@@ -340,7 +91,7 @@ const bloom = ({ ctx, resourceCache, descriptors, bloomLevels = 9 }) => [
   })),
   ...Array.from({ length: bloomLevels - 1 }, (_, k) => k).map((i) => ({
     name: `main[${i}]`,
-    frag: SHADERS.upsample.frag,
+    frag: postprocessingShaders.upsample.frag,
     // prettier-ignore
     flagDefinitions: [
       [["postProcessing", "bloom"], "USE_UPSAMPLE"],
@@ -355,7 +106,7 @@ const bloom = ({ ctx, resourceCache, descriptors, bloomLevels = 9 }) => [
 const final = () => [
   {
     name: "main",
-    frag: SHADERS.postProcessing.frag.replace(
+    frag: postprocessingShaders.postProcessing.frag.replace(
       /*glsl*/ `vec4 ssao(vec4 color, vec4 aoData, float intensity) {
   #ifdef USE_SSAO_COLORS
     vec3 rgb = mix(color.rgb, color.rgb * gtaoMultiBounce(aoData.a, color.rgb), intensity);
